# PERF: Improve performance of nanmin/nanmax for axis=None