Improve `nanmin` and `nanmax` performance primarily for smaller dtypes (`int32`, `float32`)
```
$ asv compare upstream/master HEAD --only-changed -s --sort ratio
       before           after         ratio
     [70591905]       [36bcd6e8]
     <master~1>       <nanmin>  
-        892±20μs        697±200μs     0.78  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-     1.21±0.05μs         909±50ns     0.75  reduce.Time1DReductions.time_nanmin('float32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        892±40μs        626±200μs     0.70  reduce.Time2DReductions.time_nanmin('float32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        8.86±3ms         6.05±2ms     0.68  reduce.Time1DReductions.time_nanmin('float32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        87.8±3μs         59.1±3μs     0.67  reduce.Time1DReductions.time_nanmin('float32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        930±50ns         568±30ns     0.61  reduce.Time1DReductions.time_nanmin('int32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-      5.95±0.4ms       2.85±0.2ms     0.48  reduce.Time1DReductions.time_nanmin('int32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        600±40μs         277±20μs     0.46  reduce.Time2DReductions.time_nanmin('int32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        594±10μs         260±30μs     0.44  reduce.Time2DReductions.time_nanmin('int32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
-        58.5±2μs       23.4±0.7μs     0.40  reduce.Time1DReductions.time_nanmin('int32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-CXXg++]
```