@qwhelan 2020-01-01
Implemented in #298